OBGM-47 Refactor InventoryExcelImporter - create InventoryImportDataService with validation and import methods
